utils: reorganize SimpleSet definition
